Implementations of basic seafloor detection in mask subpackage; example notebook coming soon.
I have tested two possible ways to integrate the seafloor detection into the current workflow. Both functions are currently implemented in the mask subpackage (the second one is temporarily located in the seafloor_detection subpackage within mask).

An example notebook demonstrating and testing the feature will be submitted soon to the echopype-examples repository for review.